Trade done

+++ Views/TradeView.py
@@ -20,7 +20,9 @@ class TradeView(discord.ui.View):
 		self.userdata = userData
 		self.targetdata = targetData
 		self.initial = True
-		super().__init__(timeout=30)
+		self.message = None
+		self.trademessage = None
+		super().__init__(timeout=300)
 		cnclBtn = discord.ui.Button(label="Cancel", style=discord.ButtonStyle.danger)
 		cnclBtn.callback = self.cancel_button
 		sndbtn = discord.ui.Button(label="Send", style=discord.ButtonStyle.success)
@@ -34,7 +36,10 @@ class TradeView(discord.ui.View):
 
 	async def on_timeout(self):
 		commandlockservice.DeleteLock(self.trainer.ServerId, self.trainer.UserId)
-		await self.message.delete(delay=0.1)
+		if self.message:
+			await self.message.delete(delay=0.1)
+		if self.trademessage:
+			await self.trademessage.delete(delay=0.1)
 		return await super().on_timeout()
 
 	async def PokemonSelection(self, inter: discord.Interaction, choice: str):
@@ -45,27 +50,23 @@ class TradeView(discord.ui.View):
 
 	@defer
 	async def cancel_button(self, inter: discord.Interaction):
+		self.stop()
+		commandlockservice.DeleteLock(self.trainer.ServerId, self.trainer.UserId)
 		if inter.user.id == self.targettrainer.UserId:
-			await self.message.edit(content=f'<@{self.trainer.UserId}>\nTrade offer was rejected.', embed=None, view=None)
+			await self.trademessage.edit(content=f'<@{self.trainer.UserId}>\nTrade offer was rejected.', embed=None, view=None)
+		elif self.trademessage:
+			await self.trademessage.edit(content='Trade was canceled.', embed=None, view=None)
 		else:
 			await self.message.edit(content='Trade was canceled.', embed=None, view=None)
-		commandlockservice.DeleteLock(self.trainer.ServerId, self.trainer.UserId)
 
 
-	@defer
 	async def submit_button(self, inter: discord.Interaction):
+		if not self.userpkmnchoice or not self.targetpkmnchoice:
+			return await inter.response.defer()
+		await inter.response.defer(thinking=(self.trademessage is None))
+
 		if self.initial:
-			if not self.userpkmnchoice or not self.targetpkmnchoice or inter.user.id != self.trainer.UserId:
-				return
-			
 			self.clear_items()
-			self.targettrainer = trainerservice.GetTrainer(self.targettrainer.ServerId, self.targettrainer.UserId)
-			self.targetpkmnchoice = next((p for p in self.targettrainer.OwnedPokemon if p.Id == self.targetpkmnchoice.Id and p.Pokemon_Id == self.targetpkmnchoice.Pokemon_Id), None)
-			if not self.targettrainer or not self.targetpkmnchoice:
-				commandlockservice.DeleteLock(self.trainer.ServerId, self.trainer.UserId)
-				self.stop()
-				return await self.message.edit(content=f'The target trainer no longer owns the desired Pokemon. Try again.', embed=None, view=None)
-		
 			await self.message.delete(delay=0.01)
 			rjctbtn = discord.ui.Button(label="Reject", style=discord.ButtonStyle.danger)
 			rjctbtn.callback = self.cancel_button
@@ -80,19 +81,25 @@ class TradeView(discord.ui.View):
 				f'__You Give__:\n{self.PrintPkmnDetails(self.targetpkmnchoice, self.targetdata)}\n\n__You Receive__:\n{self.PrintPkmnDetails(self.userpkmnchoice, self.userdata)}', 
 				TradeColor,
 				thumbnail=botImage)
-			await inter.followup.send(content=f'<@{self.targettrainer.UserId}>', embed=embed, view=self, ephemeral=False)
-			self.message = await inter.original_response()
+			await inter.followup.send(content=f'<@{self.targettrainer.UserId}>', embed=embed, view=self)
+			self.trademessage = await inter.original_response()
 		else:
 			if inter.user.id != self.targettrainer.UserId:
 				return
+			if commandlockservice.IsEliteFourLocked(self.targettrainer.ServerId, self.targettrainer.UserId):
+				commandlockservice.DeleteLock(self.trainer.ServerId, self.trainer.UserId)
+				self.stop()
+				return await self.trademessage.edit(content=f'One user is currently taking on the Elite Four! Try again later.', embed=None, view=None)
+			if commandlockservice.IsLocked(self.targettrainer.ServerId, self.targettrainer.UserId):
+				return await self.trademessage.edit(content=f'One user is currently in an interaction! Finish the previous command and try again.')
 			self.targettrainer = trainerservice.GetTrainer(self.targettrainer.ServerId, self.targettrainer.UserId)
 			self.targetpkmnchoice = next((p for p in self.targettrainer.OwnedPokemon if p.Id == self.targetpkmnchoice.Id and p.Pokemon_Id == self.targetpkmnchoice.Pokemon_Id), None)
 			if not self.targetpkmnchoice:
-				return await self.message.edit(content=f'The Pokemon being traded no longer exists. Please try again.', embed=None, view=None)
+				return await self.trademessage.edit(content=f'The Pokemon being traded no longer exists. Please try again.', embed=None, view=None)
 			trainerservice.TradePokemon(self.trainer, self.userpkmnchoice, self.targettrainer, self.targetpkmnchoice)
 			commandlockservice.DeleteLock(self.trainer.ServerId, self.trainer.UserId)
 			self.stop()
-			return await self.message.edit(content=f'<@{self.trainer.UserId}> traded away **{pokemonservice.GetPokemonDisplayName(self.userpkmnchoice, self.userdata)}** to <@{self.targettrainer.UserId}> for **{pokemonservice.GetPokemonDisplayName(self.targetpkmnchoice, self.targetdata)}**', embed=None, view=None)
+			return await self.trademessage.edit(content=f'<@{self.trainer.UserId}> traded away **{pokemonservice.GetPokemonDisplayName(self.userpkmnchoice, self.userdata)}** to <@{self.targettrainer.UserId}> for **{pokemonservice.GetPokemonDisplayName(self.targetpkmnchoice, self.targetdata)}**', embed=None, view=None)
 
 	def PrintPkmnDetails(self, pokemon: Pokemon, data: PokemonData):
 		pkmnData = t2a(
